pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova"
version = "0.1.0"
description = "Nova — a Discord bot serving as an OpenClaw Discord Channel."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Nova contributors" }]
keywords = ["discord", "bot", "openclaw", "slash-commands"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "discord.py[voice]>=2.4,<3.0",
    "PyNaCl>=1.5,<2.0",
    "yt-dlp>=2024.8.6",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "python-dotenv>=1.0,<2.0",
    "PyYAML>=6.0,<7.0",
    "aiohttp>=3.9,<4.0",
    "rich>=13.7,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=4.1,<5.0",
    "ruff>=0.5,<1.0",
    "black>=24.3,<25.0",
    "mypy>=1.10,<2.0",
    "types-PyYAML",
    "pre-commit>=3.7,<4.0",
]

[project.scripts]
# `nova` entry point — runs the bot.
nova = "nova.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nova = ["py.typed"]

# ---------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "SIM",           # flake8-simplify
    "RUF",           # ruff-specific
]
ignore = [
    "E501",          # line length handled by black
    "B008",          # function-call-default — fine for Depends-style args
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802"]

# ---------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py311", "py312"]

# ---------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_any_unimported = false
ignore_missing_imports = true
files = ["src"]

# ---------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["error"]

[tool.coverage.run]
branch = true
source = ["nova"]

[tool.coverage.report]
show_missing = true
skip_covered = true
